Added location inputs in ticket
* Installed semantic ui

yarn add semantic-ui-react
yarn add semantic-ui-css

(The styles of semantic ui will work properly only when semantic-ui-css is installed. To use it, include the css import statement at index.js)

* BUG FIX:
The counter in react-score-indicator was not displayed properly due to the css override of semantic ui. Then fixed it by adjusting the padding of the text using the prop 'textStyle={{paddingTop: 5}}'

* Added state 'stops' which serves as the option for ticket input dropdown of drop point

* Added TicketIput component